Refactor CIM ClusterDetailsFormFields
Since OCM uses it's own ClusterDetailsFormFields component, this refactor moves
original ClusterDetailsFormFields from common to cim directory. Removes unused
RH DNS service option and aligns disabled fields to use StaticTextField

Fixes BZ2040154
