+++ locales/en/translation.json
@@ -25,7 +25,6 @@
   "ai:1-54 characters": "1-54 characters",
   "ai:1-63 characters": "1-63 characters",
   "ai:2-54 characters": "2-54 characters",
-  "ai:A base domain will be provided for temporary, non-production clusters.": "A base domain will be provided for temporary, non-production clusters.",
   "ai:A change will require booting hosts with a new discovery ISO file.": "A change will require booting hosts with a new discovery ISO file.",
   "ai:A cluster can have {{HOSTS_MIN_COUNT}} hosts at minimum.": "A cluster can have {{HOSTS_MIN_COUNT}} hosts at minimum.",
   "ai:A cluster needs 2 or more worker nodes.": "A cluster needs 2 or more worker nodes.",
@@ -622,7 +621,6 @@
   "ai:Upload a YAML file with BareMetalHost and Secret resource definitions. Use the provided template as a reference.": "Upload a YAML file with BareMetalHost and Secret resource definitions. Use the provided template as a reference.",
   "ai:URL must start with <bold>http</bold>.": "URL must start with <bold>http</bold>.",
   "ai:Use a comma to separate each listed domain. Preface a domain with <bold>.</bold> to include its subdomains. Use <bold>*</bold> to bypass the proxy for all destinations.": "Use a comma to separate each listed domain. Preface a domain with <bold>.</bold> to include its subdomains. Use <bold>*</bold> to bypass the proxy for all destinations.",
-  "ai:Use a temporary 60-day domain": "Use a temporary 60-day domain",
   "ai:Use advanced networking": "Use advanced networking",
   "ai:Use alphanumberic characters, dot (.), underscore (_) or hyphen (-)": "Use alphanumberic characters, dot (.), underscore (_) or hyphen (-)",
   "ai:Use lowercase alphanumberic characters, dot (.) or hyphen (-)": "Use lowercase alphanumberic characters, dot (.) or hyphen (-)",
